+++ test/map/test2.c
@@ -4,17 +4,17 @@
 #include <hhmap.h>
 #include "test.h"
 
-void fill(HHSet *S, uint m) {
-  for (ushort i = 0; i < m; ++i) {
-    ushort x = rand();
+void fill(HHSet *S, ulong m) {
+  for (uint i = 0; i < m; ++i) {
+    uint x = rand();
     test(hhmapput(S, x, i + 1));
     test(hhmapget(S, x) == i + 1);
   }
 }
 
 int main(void) {
-  uint n = 256;
-  uint m = 150;
+  ulong n = 256;
+  ulong m = 150;
   HHSet *S = hhmapnew(n, NULL, NULL);
   fill(S, m);
   test(m == S->T->m);
