+++ test/map/test3.c
@@ -5,14 +5,14 @@
 #include <hhmap.h>
 #include "test.h"
 
-uint hash(uint x) {
-  uint k = (ushort)x;
+ulong hash(ulong x) {
+  ulong k = (uint)x;
   return k * 0x9e3779b1;
 }
 
-void fill(HHSet *S, uint m) {
-  for (ushort i = 0; i < m; ++i) {
-    ushort x = rand() % (USHRT_MAX - 1) + 1;
+void fill(HHSet *S, ulong m) {
+  for (uint i = 0; i < m; ++i) {
+    uint x = rand() % (USHRT_MAX - 1) + 1;
     test(!hhmapget(S, i));
     test(hhmapput(S, i, x));
     test(hhmapget(S, i) == x);
@@ -20,8 +20,8 @@ void fill(HHSet *S, uint m) {
 }
 
 int main(void) {
-  uint n = 256;
-  uint m = 150;
+  ulong n = 256;
+  ulong m = 150;
   HHSet *S = hhmapnew(n, hash, NULL);
   fill(S, m);
   test(m == S->T->m);
